Add files via upload
This program searches for magic squares using the brute force method. For this purpose, squares of a given size in characters are filled with numbers using a random number generator and this square is then checked to see whether it is a magic square. Whereby the largest random number can be set by the user.